Extract common logic to a mixin
