+++ Manifest.txt
@@ -494,6 +494,7 @@ lib/rubygems/text.rb
 lib/rubygems/uninstaller.rb
 lib/rubygems/uri_formatter.rb
 lib/rubygems/uri_parser.rb
+lib/rubygems/uri_parsing.rb
 lib/rubygems/user_interaction.rb
 lib/rubygems/util.rb
 lib/rubygems/util/licenses.rb
